demo7-replay/service-missionmanager/requirements.txt
kafka-python==2.0.2
pandas
numpy

